requirements.txt
dotenv
flask
langchain
langchain-community
faiss-cpu
pypdf
flask_cors
requests
aiohttp
groq
flask_sqlalchemy
flask_bcrypt
flask-jwt-extended
psycopg2-binary
# flask-socketio 
eventlet
flask_apscheduler
websockets